Moves debugger menu into a separate action group.
