+++ stackdriver-debugger/resources/messages/StackdriverDebuggerBundle.properties
@@ -91,5 +91,6 @@ clouddebug.runconfig.formoredetails=For more details,
 clouddebug.runconfig.documentation.url=https://cloud.google.com/debugger/
 clouddebug.runconfig.documentation.url.text=click here
 
+clouddebug.general.menu.group.text=Stackdriver Debugger
 appengine.tools.menu.debug.text=Attach Stackdriver Debugger
 appengine.tools.menu.debug.description=Create or Run a Stackdriver Debugger Configuration
